inital smart accounts
