+++ contracts/smart-accounts/btsg-eth/Readme.md
@@ -1,3 +1,3 @@
 # btsg-eth
-authentication for eth based signatures of off-chain actions.
 
+authentication for eth based signatures of off-chain actions.
