+++ schema/btsg-dao.json
@@ -1,5 +1,5 @@
 {
-  "contract_name": "btsg-eth",
+  "contract_name": "btsg-dao",
   "contract_version": "0.3.0",
   "idl_version": "1.0.0",
   "instantiate": {
@@ -7,32 +7,14 @@
     "title": "InstantiateMsg",
     "type": "object",
     "required": [
-      "pubkey"
+      "dao_addr"
     ],
     "properties": {
-      "owner": {
-        "description": "address of account to make use of this authenticator",
-        "anyOf": [
-          {
-            "$ref": "#/definitions/Addr"
-          },
-          {
-            "type": "null"
-          }
-        ]
-      },
-      "pubkey": {
-        "description": "ethereum wallet public key to sign",
+      "dao_addr": {
         "type": "string"
       }
     },
-    "additionalProperties": false,
-    "definitions": {
-      "Addr": {
-        "description": "A human readable address.\n\nIn Cosmos, this is typically bech32 encoded. But for multi-chain smart contracts no assumptions should be made other than being UTF-8 encoded and of reasonable length.\n\nThis type represents a validated address. It can be created in the following ways 1. Use `Addr::unchecked(input)` 2. Use `let checked: Addr = deps.api.addr_validate(input)?` 3. Use `let checked: Addr = deps.api.addr_humanize(canonical_addr)?` 4. Deserialize from JSON. This must only be done from JSON that was validated before such as a contract's state. `Addr` must not be used in messages sent by the user because this would result in unvalidated instances.\n\nThis type is immutable. If you really need to mutate it (Really? Are you sure?), create a mutable copy using `let mut mutable = Addr::to_string()` and operate on that `String` instance.",
-        "type": "string"
-      }
-    }
+    "additionalProperties": false
   },
   "execute": {
     "$schema": "http://json-schema.org/draft-07/schema#",
@@ -43,8 +25,21 @@
   "query": {
     "$schema": "http://json-schema.org/draft-07/schema#",
     "title": "QueryMsg",
-    "type": "string",
-    "enum": []
+    "oneOf": [
+      {
+        "type": "object",
+        "required": [
+          "get_dao"
+        ],
+        "properties": {
+          "get_dao": {
+            "type": "object",
+            "additionalProperties": false
+          }
+        },
+        "additionalProperties": false
+      }
+    ]
   },
   "migrate": null,
   "sudo": {
@@ -484,5 +479,11 @@
       }
     }
   },
-  "responses": {}
+  "responses": {
+    "get_dao": {
+      "$schema": "http://json-schema.org/draft-07/schema#",
+      "title": "String",
+      "type": "string"
+    }
+  }
 }
