+++ schema/btsg-ed25519.json
@@ -1,6 +1,6 @@
 {
   "contract_name": "btsg-ed25519",
-  "contract_version": "0.2.0",
+  "contract_version": "0.3.0",
   "idl_version": "1.0.0",
   "instantiate": {
     "$schema": "http://json-schema.org/draft-07/schema#",
